backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=4.0.0,<4.1.0",
    "fastapi>=0.121.2",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
